Create VARARG method overloading test.java
The VARARGS allow the method to accept zero or multiple arguments. before VARARGS either we use overloaded method or take an array as method parameter but it was not considered good because it leads to the maintenance problem. if we will have to pass in method, VARARGS is the better approach.
VARARGS get least priority. i.e. if no other method matched then only VARARGS method will get chance, because INT came in 1.0 version and VARARGS came in version1.3.