Create Method over riding test.java
Method over riding allow a sub class or child class to provide a specific implementation of a method that is already provided by one of it's supper classes or parent classes. The implementation in the sub class overrides (replace) the implementation in the supper class by providing a method that has same name, same return type as the method in the parent class.