Create methods can override test.java
Final method can't be override. if we don't want a method to be overridden, we declare it is final.
Static method can't overridden (Method overriding vs method hiding). When you define a static method with same signature as a static method in base class, it is known as method hiding.
Private methods can't be overridden. private methods can't be overridden as they are bonded during compile time, therefore we can't even override private methods in sub-class.